backend/requirements.txt
fastapi
uvicorn
python-socketio
python-dotenv
redis
motor
pydantic
httpx
yt-dlp
ytmusicapi
